+++ src/v1/MediaType/mustMatchContentType.ts
@@ -33,10 +33,9 @@
 //
 import { OnError, THROW_THE_ERROR } from "@ganbarodigital/ts-lib-error-reporting/lib/v1";
 
-import { MediaType } from "./MediaType";
 import { UnexpectedContentTypeError } from "../Errors/UnexpectedContentType";
 import { matchesContentType } from "./matchesContentType";
-import { contentTypeFromMediaType } from "../ContentType";
+import { resolveToContentType, ContentTypeOrMediaType } from "../Helpers";
 
 /**
  * Data guarantee. Calls your onError handler if the given input
@@ -44,7 +43,11 @@ import { contentTypeFromMediaType } from "../ContentType";
  *
  * We compare everything except the parameters of the MediaTypes.
  */
-export function mustMatchContentType(input: MediaType, safelist: MediaType[], onError: OnError = THROW_THE_ERROR): void {
+export function mustMatchContentType(
+    input: ContentTypeOrMediaType,
+    safelist: ContentTypeOrMediaType[],
+    onError: OnError = THROW_THE_ERROR,
+): void {
     // does it match?
     if (matchesContentType(input, safelist)) {
         // yes it does!
@@ -53,13 +56,13 @@ export function mustMatchContentType(input: MediaType, safelist: MediaType[], on
 
     // which content types have we checked it against?
     const checkedAgainst = safelist.map(
-        (mt) => contentTypeFromMediaType(mt)
+        (mt) => resolveToContentType(mt)
     );
 
     // tell the caller what happened
     onError(new UnexpectedContentTypeError({
         public: {
-            input: contentTypeFromMediaType(input),
+            input: resolveToContentType(input),
             required: {
                 anyOf: checkedAgainst,
             },