+++ CHANGELOG.md
@@ -30,6 +30,8 @@ The following changes have been completed, and will be included in the next tagg
   - added `isContentType()` data guard
   - added `mustBeContentType()` data guarantee
   - added `NotAContentTypeError` throwable error
+* Errors
+  - UnexpectedContentTypeError now requires a `ContentType` as the public data
 
 ## v1.2.1